Added a card that displays all information about the current change
