+++ src/App.tsx
@@ -4,7 +4,7 @@ import getWeb3 from "./util/getWeb3";
 
 const appStyles = require("./App.css");
 
-import Proposer from "./components/Proposer";
+import ChangeCard from "./components/ChangeCard";
 
 interface IAppState {
   web3: Web3;
@@ -28,26 +28,7 @@ class App extends React.Component<{}, IAppState> {
   public render() {
     return (
       <div className={appStyles.app}>
-        <div className={appStyles.appHeader}>
-          <h2>React Ethereum DApp Template</h2>
-        </div>
-        <div className={appStyles.appIntro}>
-          {this.state.web3 ? (
-            <div>
-              <p>
-                Provider is MetaMask: {(this.state.web3.currentProvider as any).isMetaMask ? "yes" : "no"}
-              </p>
-              <p>
-                Provider is Mist: {(window as any).mist ? "yes" : "no"}
-              </p>
-              {(this.state.web3.currentProvider as any).host ?
-                <p>Provider is {(this.state.web3.currentProvider as any).host}</p> : null}
-            </div>
-          ) :
-            <p>Web3 is loading</p>}
-        </div>
-        <hr />
-        {this.state.web3 ? <Proposer web3={this.state.web3} /> : null}
+        {this.state.web3 ? <ChangeCard web3={this.state.web3} /> : null}
       </div>
     );
   }
